* Refactoring
